# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DCP Platform API
  version: 5.1.0
  description: |
    DCP is Saudi Arabia's sovereign AI compute platform — built to be used by AI
    agents and software, not only humans. It exposes:

    - **OpenAI-compatible inference** at `https://api.dcp.sa/v1`
      (`POST /v1/chat/completions`, `GET /v1/models`) — a drop-in replacement
      for `api.openai.com`. Set `base_url` and a renter Bearer key; no rewrite.
    - **On-demand GPU rental (pods)** — rent a whole NVIDIA GPU (H200, H100,
      A100, RTX 4090 and more) with root + Jupyter + SSH in about a minute via
      `POST /api/pods`. Billed prepaid per GPU-second in SAR.
    - **Persistent in-Kingdom storage (volumes)** via `POST /api/volumes/rent`.
    - An official **MCP server** (`integrations/dcp-mcp`) exposing all of the
      above as Model Context Protocol tools for agents.

    All compute runs on Saudi-owned hardware inside the Kingdom under full PDPL
    data-residency compliance. Data, models, and storage never leave Saudi
    Arabia unless a tenant explicitly opts in to cross-border frontier models.

    The platform is also a two-sided marketplace: providers register NVIDIA
    GPUs, install the daemon, and earn SAR; renters consume inference, pods, and
    volumes. Money-mutating writes accept an `Idempotency-Key` header so agent
    retries are safe, and return HTTP 402 with a structured payment-required
    body when the prepaid wallet cannot cover a request.

    ## Currency
...

...

...

...

...
servers:
...
components:
  securitySchemes:
    ProviderApiKey:
      type: apiKey
      in: query
      name: key
      description: Provider API key. Prefix `dc1-provider-`. Also accepted as `x-provider-key` header.
    RenterApiKey:
      type: apiKey
      in: query
      name: key
      description: "Renter API key. Master keys use prefix `dcp-renter-`; sub-keys (minted via /api/renters/me/keys) use `dc1-sk-`; legacy `dc1-renter-` also accepted. Pass as `?key=`, the `x-renter-key` header, or `Authorization: Bearer <key>`."
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin token set via DC1_ADMIN_TOKEN env var. Also accepted as Bearer token.
  schemas:
...
    BatchInferenceJob:
      type: object
      description: Batch inference metadata record for validated JSONL input. In this foundation slice, execution remains disabled until the worker/result/billing rail is implemented.
      required: [batch_id, renter_id, status, input_storage_key, input_checksum_sha256, request_count, completion_window, execution_enabled, results_available, created_at, updated_at]
      properties:
        batch_id:
          type: string
          example: batch_8f4c2b9a18e34d220f7a9b12
        renter_id:
          type: integer
          example: 42
        status:
          type: string
          enum: [created, queued, running, completed, failed, cancelled]
          example: created
        input_storage_key:
          type: string
          description: Relative object key reserved for the normalized JSONL input artifact.
          example: batch-inputs/renter-42/batch_8f4c2b9a18e34d220f7a9b12/input.jsonl
        input_checksum_sha256:
          type: string
          description: SHA-256 digest of the normalized JSONL input.
          example: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
        input_normalized_bytes:
          type: integer
          example: 264
        request_count:
          type: integer
          example: 2
        completion_window:
          type: string
          enum: [24h]
          example: 24h
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        result_storage_key:
          type: string
          nullable: true
          example: batch-results/renter-42/batch_8f4c2b9a18e34d220f7a9b12/output.jsonl
        result_checksum_sha256:
          type: string
          nullable: true
          description: SHA-256 digest of the normalized result JSONL artifact. Required before results are considered available.
          example: bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
        result_normalized_bytes:
          type: integer
          example: 2048
        completed_count:
          type: integer
          example: 0
        failed_count:
          type: integer
          example: 0
        total_cost_halala:
          type: integer
          example: 0
        idempotency_key:
          type: string
          nullable: true
          example: batch-create-2026-07-08
        execution_enabled:
          type: boolean
          example: false
        results_available:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    BatchInferenceResultManifest:
      type: object
      description: Read-only result artifact manifest for one batch. It exposes proof metadata and can include a short-lived signed download URL when result object-store signing is configured.
      required: [batch_id, renter_id, status, results_available, download_enabled, next]
      properties:
        batch_id:
          type: string
          example: batch_8f4c2b9a18e34d220f7a9b12
        renter_id:
          type: integer
          example: 42
        status:
          type: string
          enum: [created, queued, running, completed, failed, cancelled]
          example: completed
        results_available:
          type: boolean
          description: True only when the batch is completed and has both a result storage key and SHA-256 checksum.
          example: true
        result_storage_key:
          type: string
          nullable: true
          example: batch-results/renter-42/batch_8f4c2b9a18e34d220f7a9b12/output.jsonl
        result_checksum_sha256:
          type: string
          nullable: true
          example: bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
        result_normalized_bytes:
          type: integer
          example: 2048
        completed_count:
          type: integer
          example: 2
        failed_count:
          type: integer
          example: 0
        total_cost_halala:
          type: integer
          example: 18
        download_enabled:
          type: boolean
          example: false
        download_url:
          type: string
          nullable: true
          description: Short-lived object-store download URL. Present only when results are complete, checksum proof exists, the storage key is scoped to the renter/batch, and batch result object-store signing is configured.
        download_method:
          type: string
          nullable: true
          enum: [GET]
          example: GET
        download_expires_in:
          type: integer
          nullable: true
          description: Signed URL lifetime in seconds.
          example: 900
        download_expires_at:
          type: string
          format: date-time
          nullable: true
        download_configured:
          type: boolean
          nullable: true
          description: False when result proof exists but `BATCH_RESULTS_S3_*` object-store signing is not configured.
        next:
          type: string
          example: configure_batch_result_object_store
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    BatchInferenceLine:
      type: object
      description: Per-line batch inference ledger row. Raw prompt/request bodies are not returned; callers get custom id, endpoint, model, checksums, lifecycle, usage, and cost metadata.
      required: [batch_id, renter_id, line_index, custom_id, method, url, model_id, request_checksum_sha256, status, usage, cost_halala, created_at, updated_at]
      properties:
        batch_id:
          type: string
          example: batch_8f4c2b9a18e34d220f7a9b12
        renter_id:
          type: integer
          example: 42
        line_index:
          type: integer
          example: 1
        custom_id:
          type: string
          example: request-1
        method:
          type: string
          enum: [POST]
          example: POST
        url:
          type: string
          enum: [/v1/chat/completions, /v1/complete]
          example: /v1/chat/completions
        model_id:
          type: string
          example: qwen/qwen3-coder
        request_checksum_sha256:
          type: string
          description: SHA-256 digest of the normalized batch request line.
          example: cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
          example: pending
        status_code:
          type: integer
          nullable: true
          example: 200
        response_checksum_sha256:
          type: string
          nullable: true
          example: dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd
        response_normalized_bytes:
          type: integer
          example: 0
        usage:
          type: object
          required: [prompt_tokens, completion_tokens, total_tokens]
          properties:
            prompt_tokens:
              type: integer
              example: 0
            completion_tokens:
              type: integer
              example: 0
            total_tokens:
              type: integer
              example: 0
        cost_halala:
          type: integer
          example: 0
        request_id:
          type: string
          nullable: true
          example: batch_8f4c2b9a18e34d220f7a9b12:request-1
        provider_response_id:
          type: string
          nullable: true
        error_code:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
...
paths:
...
  /api/batches:
    get:
      tags: [Batches]
      summary: List renter batch inference records
      description: |
        Returns validated batch-inference metadata records for the authenticated
        renter. This foundation route does not execute inference, produce result
        artifacts, apply batch discounts, or flip `/v1/models` batch capability
        flags yet.
      security:
        - RenterApiKey: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [created, queued, running, completed, failed, cancelled]
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - in: query
          name: offset
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Batch record list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: { type: string, example: list }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/BatchInferenceJob' }
                  count: { type: integer, example: 1 }
                  limit: { type: integer, example: 50 }
                  offset: { type: integer, example: 0 }
    post:
      tags: [Batches]
      summary: Validate JSONL and create a batch inference metadata record
      description: |
        Validates OpenAI-style JSONL lines for `/v1/chat/completions` and
        `/v1/complete`, records checksum/count metadata, and returns a stable
        batch id. Execution remains disabled until the worker/result/billing
        slice lands.
      security:
        - RenterApiKey: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
            maxLength: 180
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input_jsonl]
              properties:
                batch_id:
                  type: string
                  description: Optional client-stable id. Generated when omitted.
                  example: batch_8f4c2b9a18e34d220f7a9b12
                input_jsonl:
                  type: string
                  description: Newline-delimited JSON batch requests.
                  example: '{"custom_id":"req-1","method":"POST","url":"/v1/chat/completions","body":{"model":"qwen/qwen3-coder","messages":[{"role":"user","content":"hello"}]}}'
                input_storage_key:
                  type: string
                  description: Optional relative object key reserved for the normalized input artifact.
                completion_window:
                  type: string
                  enum: [24h]
                  default: 24h
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Batch metadata record created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  batch: { $ref: '#/components/schemas/BatchInferenceJob' }
                  idempotent_replay:
                    type: boolean
                    example: false
                  execution_enabled:
                    type: boolean
                    example: false
                  next:
                    type: string
                    example: batch_worker_and_result_artifact_not_enabled
        '400':
          description: Invalid batch JSONL or metadata.
        '409':
          description: Batch id already exists.
  /api/batches/{batch_id}:
    get:
      tags: [Batches]
      summary: Get one renter batch inference record
      security:
        - RenterApiKey: []
      parameters:
        - in: path
          name: batch_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Batch detail.
          content:
            application/json:
              schema:
                type: object
                properties:
                  batch: { $ref: '#/components/schemas/BatchInferenceJob' }
        '404':
          description: Batch not found for this renter.
  /api/batches/{batch_id}/lines:
    get:
      tags: [Batches]
      summary: List one renter batch's per-line ledger rows
      description: |
        Returns per-line batch status and accounting metadata without returning
        raw prompts, messages, or completion bodies. The line ledger is created
        at batch submission time and is the future write target for worker
        execution, usage, and settlement proof.
      security:
        - RenterApiKey: []
      parameters:
        - in: path
          name: batch_id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Batch line ledger.
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    enum: [list]
                  batch_id:
                    type: string
                  status:
                    type: string
                    enum: [created, queued, running, completed, failed, cancelled]
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/BatchInferenceLine' }
                  count:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '404':
          description: Batch not found for this renter.
  /api/batches/{batch_id}/results:
    get:
      tags: [Batches]
      summary: Get one renter batch result manifest
      description: |
        Returns result artifact metadata for a batch. Results are considered
        available only after the batch is completed with both a result storage
        key and SHA-256 checksum. When `BATCH_RESULTS_S3_BUCKET` plus S3
        endpoint/key/secret configuration is present, the route includes a
        short-lived signed GET URL scoped to the completed result object.
        This route does not enable production batch execution or batch
        discounts by itself.
      security:
        - RenterApiKey: []
      parameters:
        - in: path
          name: batch_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Batch result manifest.
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { $ref: '#/components/schemas/BatchInferenceResultManifest' }
        '404':
          description: Batch not found for this renter.
...